Fix circular deps. Comment code.
